Add files via upload
Hello: The reason this is uploaded late is because I just realized what the issue with the PDF formatting was. Apparently something is wrong with my code where /newpage does not work and completely breaks the knit. That's the only change between this version and my RMD that was uploaded last week, so I hope this doesn't count as late. If it does, please just grade what I had turned in at the time. Thank you!